tests for openshift-workload-reload celery task (#79)
Tests are built calling the task synchronously so that we test the whole
celery setup without the need of the broker or the worker, being the
test case and the real world very close.

I have removed the reload messages from the custom base class because
it is intended to be generic. This introduces the need of keeping
somewhere else the details of the invocation, in this case the cluster,
namespace, kind and name of the workload restarted.

Signed-off-by: Rafa Porres Molina <rporresm@redhat.com>